refactor backend layout and services
Move API/search and indexer/utils under backend, add services modules, and update config path usage and docs.
