+++ backend/__init__.py
@@ -1,0 +1,1 @@
+"""Backend package for the search service."""
