+++ backend/indexer/__init__.py
@@ -1,0 +1,1 @@
+"""Indexing utilities for backend."""
