--- utils/__init__.py
@@ -1,3 +1,0 @@
-from utils.clients import get_groq_client, get_helix_client
-
-__all__ = ["get_helix_client", "get_groq_client"]
